metaspace/engine/requirements.txt
pyImagingMSpec==0.1.4
cpyImagingMSpec==0.2.4
pyMSpec==0.1.2
cpyMSpec==0.4.2  # Drop all isotopic patterns if updated!
cpyMSpec_0_3_5==0.3.5
pyimzML==1.4.1
requests==2.26.0
pytest
boto3==1.17.13
fabric3
pypng==0.0.19    # 0.0.20 introduced incompatible API changes
pyyaml>=5.1
elasticsearch-dsl>=5.0.0,<6.0.0
pika==0.13.1
bottle
Pillow==8.2.0
numpy==1.19.5
scipy==1.6.0
pandas==1.2.0
cffi
psycopg2-binary==2.8.6
cherrypy<9
redis
scikit-learn==0.23.2
pyspark[sql]==3.0.1
pyarrow==1.0.1
pysparkling==0.6.0
s3fs
lithops==2.5.0
numba==0.54.0
catboost==1.0.3
zstandard==0.15.2